% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{bma}
\alias{bma}
\title{Bayesian Model Averaging}
\usage{
bma(object, top = 2, ...)
}
\arguments{
\item{object}{\code{melsm} object}

\item{top}{number of models to average}

\item{...}{currently ignored}
}
\value{
object of class \code{melsm}
}
\description{
Bayesian Model Averaging
}
\examples{
\dontrun{
dat <- nlme::Orthodont
fit <- melsm(fixed_location = distance ~  age,
            random_location = ~ age | Subject,
            fixed_scale = sigma ~  1, k = 3,
            random_scale = ~ 1 | Subject,
            adapt = 5000,
            iter = 10000,
            rho_test = "all",
            mixture = "SSVS",
            data = dat)

# Bayesian model averaging
bma(fit, top = 2)
}
}
