% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.melsm.R
\name{coef.melsm}
\alias{coef.melsm}
\title{Extract \code{melsm} Coefficients}
\usage{
\method{coef}{melsm}(object, cred = 0.95, ...)
}
\arguments{
\item{object}{\code{melsm} object}

\item{cred}{credible interval}

\item{...}{currently ignored}
}
\value{
3D array of class \code{coef.melsm}
}
\description{
Extract coefficients for each cluster (e.g., person, etc.), which obtained by adding the fixed effect
to each random effect.
}
\note{
this function differs from \code{nlme},
in that additional information (not only the coefficients) is returned in a list. This is used
to plot the coefficients (\code{\link{plot.coef.melsm}})
}
\examples{
\dontrun{
dat <- nlme::Orthodont
fit <- melsm(fixed_location = distance ~  age,
            random_location = ~ age | Subject,
            fixed_scale = sigma ~  1, k = 3,
            random_scale = ~ 1 | Subject,
            adapt = 5000,
            iter = 10000,
            rho_test = "all",
            mixture = "SSVS",
            data = dat)

coef(fit, cred = 0.90)
}
}
