% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm.R
\name{confirm.melsm}
\alias{confirm.melsm}
\title{Confirmatory Hypothesis Testing for \code{melsm} Objects}
\usage{
\method{confirm}{melsm}(object, hyp, ...)
}
\arguments{
\item{object}{object of class \code{melsm}}

\item{hyp}{hypothesis to test}

\item{...}{currently ignored}
}
\value{
object of class \code{confirm}
\itemize{
\item \code{BF} Bayes factor
\item \code{post_probs} posterior probabilities
\item \code{prior_probs} prior probabilities
\item \code{model} model type (\code{melsm})
\item \code{mixture} mixture type
}
}
\description{
Confirmatory hypotheses testing of scientific expectations.
}
\examples{

\dontrun{
dat <- nlme::Orthodont
fit <- melsm(fixed_location = distance ~  age,
            random_location = ~ age | Subject,
            fixed_scale = sigma ~  1, k = 3,
            random_scale = ~ 1 | Subject,
            adapt = 5000,
            iter = 10000,
            rho_test = "all",
            mixture = "SSVS",
            data = dat)


# hypothesis (rho_01 is null and rho_02 is negative )
# vs compliment (not h1)
hyp <- list(h1 = c("rho_01 = k1", "rho_02 = k3"))
confirm(fit, hyp)

# hypothesis  (both positive) vs both null
hyp <- list(h1 = c("rho_01 = k2", "rho_02 = k2"),
h2 = c("rho_01 = k1", "rho_02 = k1"))
confirm(fit, hyp)

}
}
