% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm.R
\name{confirm}
\alias{confirm}
\title{S3 \code{confirm} method}
\usage{
confirm(object, hyp, ...)
}
\arguments{
\item{object}{\code{melsm} object}

\item{hyp}{hypothesis to test}

\item{...}{currently ignored}
}
\description{
S3 \code{confirm} method
}
\seealso{
\code{\link{confirm}}
}
