% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_plot.R
\name{cor_plot}
\alias{cor_plot}
\title{Plot Coefficient Scatteplots}
\usage{
cor_plot(object, palette = "YlOrRd", direction = -1, ...)
}
\arguments{
\item{object}{\code{melsm} object}

\item{palette}{paletter for 2d density heat map}

\item{direction}{direction of the gradient}

\item{...}{currently ignored}
}
\value{
list of \code{ggplot} objects
}
\description{
Plot Coefficient Scatteplots
}
\examples{



\dontrun{
fit_int <- melsm(fixed_location = rt ~ 1,
                 random_location = ~ 1| ID,
                 fixed_scale = sigma ~ 1,
                 random_scale = ~ 1 | ID, k = 3,
                 mixture = "SSVS", adapt = 5000,
                 iter = 10000, data = stroop)

cor_plot(fit_int)
}
}
