% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_bf.R
\name{marginal_bf}
\alias{marginal_bf}
\title{Compute Marginal Bayes Factors for \code{melsm} Objects}
\usage{
marginal_bf(object, H1 = "k1", H2 = "compliment", ...)
}
\arguments{
\item{object}{\code{melsm} object}

\item{H1}{hypothesis one (only for \code{mixture = "SSVS"})}

\item{H2}{hypothesis two (only for \code{mixture = "SSVS"})}

\item{...}{currently ignored}
}
\value{
list of class \code{marginal bf}
}
\description{
Compute Marginal Bayes Factors for \code{melsm} Objects
}
