% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_code.R
\name{model_code}
\alias{model_code}
\title{JAGS Model Generated by HypMuVar}
\usage{
model_code(object)
}
\arguments{
\item{object}{\code{melsm} object}
}
\value{
printed JAGS model
}
\description{
JAGS Model Generated by HypMuVar
}
\examples{
\dontrun{
dat <- nlme::Orthodont
fit <- melsm(fixed_location = distance ~  age,
            random_location = ~ age | Subject,
            fixed_scale = sigma ~  1, k = 3,
            random_scale = ~ 1 | Subject,
            adapt = 5000,
            iter = 10000,
            rho_test = "all",
            mixture = "SSVS",
            data = dat)

model_code(fit)
}
}
