% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prob.R
\name{model_prob}
\alias{model_prob}
\title{Compute Model Probabilities for \code{melsm} Objects}
\usage{
model_prob(object, ...)
}
\arguments{
\item{object}{\code{melsm} object}

\item{...}{currently ignored}
}
\value{
list of class \code{model_prob}
}
\description{
Compute Model Probabilities for \code{melsm} Objects
}
\examples{
\dontrun{
dat <- nlme::Orthodont
fit <- melsm(fixed_location = distance ~  age,
            random_location = ~ age | Subject,
            fixed_scale = sigma ~  1, k = 3,
            random_scale = ~ 1 | Subject,
            adapt = 5000,
            iter = 10000,
            rho_test = "all",
            mixture = "SSVS",
            data = dat)

model_prob(fit)
}
}
