% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_select.melsm.R
\name{model_select.melsm}
\alias{model_select.melsm}
\title{Select Model for \code{melsm} Objects}
\usage{
\method{model_select}{melsm}(object, type = "hpm", ...)
}
\arguments{
\item{object}{\code{melsm} object}

\item{type}{model to select (currently the only option is the highest posterior model \code{hpm})}

\item{...}{currently ignored}
}
\value{
object of class \code{model_select.melsm}
}
\description{
Model selection for mixed effects location scale models.
}
\examples{
\dontrun{
dat <- nlme::Orthodont
fit <- melsm(fixed_location = distance ~  age,
            random_location = ~ age | Subject,
            fixed_scale = sigma ~  1, k = 3,
            random_scale = ~ 1 | Subject,
            adapt = 5000,
            iter = 10000,
            rho_test = "all",
            mixture = "SSVS",
            data = dat)

# highest posterior model
hpm <- model_select(fit, type = "hpm")
summary(hpm, cred = 0.95)
}
}
