% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_select.melsm.R
\name{model_select}
\alias{model_select}
\title{S3 \code{model_select} method}
\usage{
model_select(...)
}
\arguments{
\item{...}{currently ignored}
}
\description{
S3 \code{model_select} method
}
\seealso{
\code{\link{model_select.melsm}}
}
