% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.melsm.R
\name{plot.coef.melsm}
\alias{plot.coef.melsm}
\title{Plot \code{coef.melsm} Objects}
\usage{
\method{plot}{coef.melsm}(x, ...)
}
\arguments{
\item{x}{\code{coef.melsm} object}

\item{...}{currently ignored}
}
\value{
list of \code{ggplot} objects
}
\description{
Plot \code{coef.melsm} Objects
}
\examples{

\dontrun{
dat <- nlme::Orthodont
fit <- melsm(fixed_location = distance ~  age,
            random_location = ~ age | Subject,
            fixed_scale = sigma ~  1, k = 3,
            random_scale = ~ 1 | Subject,
            adapt = 5000,
            iter = 10000,
            rho_test = "all",
            mixture = "SSVS",
            data = dat)

coefs <- coef(fit, cred = 0.90)

plot(coefs)
}
}
