% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.melsm.R
\name{plot.ranef.melsm}
\alias{plot.ranef.melsm}
\title{Plot \code{ranef.melsm} Objects}
\usage{
\method{plot}{ranef.melsm}(x, ...)
}
\arguments{
\item{x}{\code{ranef.melsm} object}

\item{...}{currently ignored}
}
\value{
list of \code{ggplot} objects
}
\description{
Plot \code{ranef.melsm} Objects
}
\examples{

\dontrun{
dat <- nlme::Orthodont
fit <- melsm(fixed_location = distance ~  age,
            random_location = ~ age | Subject,
            fixed_scale = sigma ~  1, k = 3,
            random_scale = ~ 1 | Subject,
            adapt = 5000,
            iter = 10000,
            rho_test = "all",
            mixture = "SSVS",
            data = dat)

res <- ranef(fit, cred = 0.90)

plot(res)
}
}
