% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.melsm.R
\name{print.coef.melsm}
\alias{print.coef.melsm}
\title{Print Method for \code{coef.melsm} Objects}
\usage{
\method{print}{coef.melsm}(x, ...)
}
\arguments{
\item{x}{\code{coef.melsm} object}

\item{...}{currently ignored}
}
\description{
Print Method for \code{coef.melsm} Objects
}
