% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm.R
\name{print.confirm}
\alias{print.confirm}
\title{Print Method for \code{confirm} Objects}
\usage{
\method{print}{confirm}(x, ...)
}
\arguments{
\item{x}{\code{confirm} object}

\item{...}{currently ignored}
}
\description{
Print Method for \code{confirm} Objects
}
\seealso{
\code{\link{confirm}}
}
