% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_bf.R
\name{print.marginal_bf}
\alias{print.marginal_bf}
\title{Print Method for \code{marginal_bf} objects}
\usage{
\method{print}{marginal_bf}(x, ...)
}
\arguments{
\item{x}{\code{marginal_bf} object}

\item{...}{currently ignored}
}
\description{
Print Method for \code{marginal_bf} objects
}
