% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melsm.R
\name{print.melsm}
\alias{print.melsm}
\title{Print Method for \code{melsm} objects}
\usage{
\method{print}{melsm}(x, ...)
}
\arguments{
\item{x}{\code{melsm} object}

\item{...}{currently ignored}
}
\description{
Print Method for \code{melsm} objects
}
