% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prob.R
\name{print.model_prob}
\alias{print.model_prob}
\title{Print Method for \code{model_prob} Objects}
\usage{
\method{print}{model_prob}(x, ...)
}
\arguments{
\item{x}{\code{model_prob} object}

\item{...}{currently ignored}
}
\description{
Print Method for \code{model_prob} Objects
}
