% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_select.melsm.R
\name{print.model_select.melsm}
\alias{print.model_select.melsm}
\title{Print Method for \code{model_select.melsm} objects}
\usage{
\method{print}{model_select.melsm}(x, ...)
}
\arguments{
\item{x}{\code{model_select.melsm} object}

\item{...}{currently ignored}
}
\description{
Print Method for \code{model_select.melsm} objects
}
\seealso{
\code{\link{model_select.melsm}}
}
