% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.melsm.R
\name{print.ranef.melsm}
\alias{print.ranef.melsm}
\title{Print Method for \code{ranef.melsm} Objects}
\usage{
\method{print}{ranef.melsm}(x, ...)
}
\arguments{
\item{x}{\code{ranef.melsm} object}

\item{...}{currently ignored}
}
\description{
Print Method for \code{ranef.melsm} Objects
}
