% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm.R
\name{print.summary.confirm}
\alias{print.summary.confirm}
\title{Print Method for \code{summary.confirm} Objects}
\usage{
\method{print}{summary.confirm}(x, ...)
}
\arguments{
\item{x}{\code{summary.confirm} object}

\item{...}{currently ignored}
}
\description{
Print Method for \code{summary.confirm} Objects
}
\seealso{
\code{\link{confirm}}
}
