% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_names.R
\name{prior_names}
\alias{prior_names}
\title{Prior Names}
\usage{
prior_names(fixed_location, random_location, fixed_scale, random_scale,
  data)
}
\arguments{
\item{fixed_location}{a \code{formula} object for the fixed-effects part of the location sub-model, with the response (on the left) and
the terms (on the right) sepearted by \code{~}.}

\item{random_location}{a \code{formula} object for the random effects part of the location sub-model, with the random effect (on the left)
and the cluster (on the right) seperated by \code{~}.}

\item{fixed_scale}{a \code{formula} object for the fixed-effects part of the scale sub-model, with the response (on the left) and
the terms (on the right) sepearted by \code{~}.}

\item{random_scale}{a \code{formula} object for the random effects part of the location sub-model, with the random effect (on the left)
and the cluster (on the right) seperated by \code{~}.}

\item{data}{a data frame  containing the variables named in \code{fixed_location}, \code{random_location},
\code{fixed_scale}, and \code{random_scale}}
}
\description{
Prior Names
}
\examples{

dat <- nlme::Orthodont
prior_names(fixed_location = distance ~ Sex + age,
           fixed_scale = sigma ~ age,
           random_location = ~ 1 | Subject,
           random_scale = ~ 1 | Subject,
           data = dat)
}
