% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.melsm.R
\name{ranef.melsm}
\alias{ranef.melsm}
\title{Extract \code{melsm} Random Effects}
\usage{
\method{ranef}{melsm}(object, cred = 0.95, ...)
}
\arguments{
\item{object}{\code{melsm} object}

\item{cred}{credible interval}

\item{...}{currently ignored}
}
\value{
3D array of class \code{ranef.melsm}
}
\description{
Extract \code{melsm} Random Effects
}
\note{
this function differs from \code{nlme},
in that additional information (not only the random effects) is returned in a list. This is used
to plot the random effects (\code{\link{plot.ranef.melsm}})
}
\examples{
\dontrun{
dat <- nlme::Orthodont
fit <- melsm(fixed_location = distance ~  age,
            random_location = ~ age | Subject,
            fixed_scale = sigma ~  1, k = 3,
            random_scale = ~ 1 | Subject,
            adapt = 5000,
            iter = 10000,
            rho_test = "all",
            mixture = "SSVS",
            data = dat)

ranef(fit, cred = 0.90)
}
}
