% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{stroop}
\alias{stroop}
\title{Stroop Task Data}
\format{A data frame with 11245 rows and 3 columns}
\usage{
data("stroop")
}
\description{
A Stroop task dataset that includes 121 subjects with roughly 90 repeated measurments each.
}
\details{
\itemize{
  \item \code{rt} reaction times on the seconds scale
  \item \code{congruency} experimental condition (congruent and incongruent primes)
  \item \code{ID} study subject
}
}
\references{
von Bastian, C. C., Souza, A. S., & Gade, M. (2016, 2). No evidence for bilingual cognitive
advantages: A test of four hypotheses. Journal of experimental psychology. General, 145 (2), 246–258. doi: 10.1037/xge0000120
}
\keyword{datasets}
