% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm.R
\name{summary.confirm}
\alias{summary.confirm}
\title{Summary Method for \code{confirm} Object}
\usage{
\method{summary}{confirm}(object, ...)
}
\arguments{
\item{object}{\code{confirm} Object}

\item{...}{currently ignored}
}
\value{
object of class \code{summary.confirm}
\itemize{
\item \code{dat_prob} data frame with probabilities
\item \code{BF_12} Bayes factor in favor of H1
\item \code{object} \code{melsm} object
}
}
\description{
Summary Method for \code{confirm} Object
}
\seealso{
\code{\link{confirm}}
}
