% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melsm.R
\name{summary.melsm}
\alias{summary.melsm}
\title{Summary Method for \code{melsm} Objects}
\usage{
\method{summary}{melsm}(object, cred = 0.95, ...)
}
\arguments{
\item{object}{\code{melsm} object}

\item{cred}{credible interval}

\item{...}{currently ignored}
}
\description{
Summary Method for \code{melsm} Objects
}
\examples{

\dontrun{
dat <- nlme::Orthodont
fit <- melsm(fixed_location = distance ~  age,
            random_location = ~ age | Subject,
            fixed_scale = sigma ~  1, k = 3,
            random_scale = ~ 1 | Subject,
            adapt = 5000,
            iter = 10000,
            rho_test = "all",
            mixture = "SSVS",
            data = dat)

summary(fit)

}
}
