% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_select.melsm.R
\name{summary.model_select.melsm}
\alias{summary.model_select.melsm}
\title{Summary Method for \code{model_select.melsm} Objects}
\usage{
\method{summary}{model_select.melsm}(object, cred = 0.95, ...)
}
\arguments{
\item{object}{\code{model_select.melsm} object}

\item{cred}{credible interval}

\item{...}{currently ignored}
}
\description{
Summary Method for \code{model_select.melsm} Objects
}
\seealso{
\code{\link{model_select.melsm}}
}
